requirements.txt
flask
requests
beautifulsoup4
openai
pinecone-client
flask-cors
python-dotenv
gunicorn
sseclient-py
langchain
langchain-community
pinecone-notebooks
tiktoken
streamlit 
tabulate
matplotlib